refactor(collection): extract DocumentManager
- Extract document CRUD, attachments and path validation to DocumentManager
- De-clutter Collection class by delegating document operations to the manager
- Centralize attachment allowlist logic
